cloud_function_source/requirements.txt
pandas
google-cloud-storage
google-cloud-bigquery
functions-framework
pyarrow